Add ability to total the prices of Cart items
A user needs to know how much their anticipated purchase will cost.

Add a "total_price" method to compute the cost of everything in the
cart.  To make this work, the Cart requires its contents to respond to
the "price" method.  We have a choice here:

* Add the PriceableStub class as the first implememtor of this implicit
  "Priceable" interface so that the test doesn't depend on
  a production Item.  This also avoids a yak shave to create a real
  Item, since we don't have one yet.  But are we over-engineering that
  interface if we don't need to support multiple Item types?  When we do
  add the Item, we'll have to create a shared example group and
  everything.

* Add the Item class and just use that in the Cart test.

Although the second choice is certainly defensible, this is an exercise
in testing in isolation.  The second choice is not in isolation; it's
more of an intergration test, as it covers integration between Item and
Cart.  So choose the first route in hopes of learning more and gaining
experience with testing in isolation.