Add an Item class with name and price
We want to have items in the cart, and they must have a price and a name
that we can use to remind the user about what they added.

Add a simple test and implementation.

One issue we've introduced is potential divergence between the Item
class and the PriceableStub, but we'll address that next.