Allow an Order to compute the amount of payment due from the user
The Order currently has no state.

Allow it to store the items from the Cart so that it knows what was
ordered.  Record the items, but not the Cart itself, as the Cart's
lifecycle ends with purchase.  Leave Order#amount_due method body as
incidental duplication of Cart#total_price for now.