Isolate the test for Order#place by adding ItemHolder interface
The Order#place method is currently being tested with real
collaborators; it's not in isolation.  That's not so bad because right
now they're simple and fast.  But it does mean the test is coupled to
those collaborators.

Replace them with a new ItemHolderStub that implements a common
interface with Cart, and place inside it a PriceableStub.