Improve Priceable and ItemHolder specs
The shared examples for ItemHolder and Priceable are not very robust.
They check for responding to required methods, but they don't check
return values.

Actually call the methods and expect a hard-coded return for Priceable
and expect an Array for ItemHolder.  Ideally we'd also examine the type
of the Array contents, but because we are duck-typing, it's hard to
invoke the ItemHolder shared example group on one of them.