Render a complete CartView including items and prices
The CartView is just a skeleton.

Fill it out to render a useful representation of Cart contents.

What interface(s) could we use to stub this test?  We cannot use
ItemHolderStub because also need Cart#total_price.  We cannot use
PriceableStub because we also need Item#name.  We could introduce
interfaces for these two roles, but they will be very specific to the
CartView: CartViewCartStub and CartViewItemStub, perhaps?  It seems
likely that CartView will need most of the state accessors of the
models.  Is there value in isolating the template?  It would allow us to
avoid creating real Cart and Items when we have many cases to test in
the template, such as empty list.