Add special case to CartView for empty Cart
When the cart is empty, the $0 price is not useful, and the "Items"
heading is confusing above an empty list.

Render a special message for an empty cart.

(Incidentally, there is a code smell here: we're switching on type.
We're asking the Cart what kind it is (empty or not) and then deciding
how to handle it in the view.  Perhaps we should let the Cart should
decide to render itself, and it should choose a different template
depending on whether it's an EmptyCart or not?  One problem is that
Carts can be mutated, so they can change between empty or not while
still being the same Cart.  That makes it hard to use polymorphism to
eliminate the conditional.)