Create stubs to test CartView in isolation
The CartView is tested with real collaborators -- Cart and Item.  This
means it will fail if one of them breaks, even if the view itself is
correct.  Also, it means potentially expensive setup for Cart and Item
just to test CartView.

Create a special interface for each collaborator that contains just the
functionality needed by CartView and which is already implemented by the
real collaborator.  Create a stub for each interface also, and test both
the stub and the real collaborator with the shared example group for the
interface.

Issues:

Due to nuance of rspec vs. MiniTest, the Cart spec runs examples from
the ItemHolder interface twice, because it's included twice, once
directly, and once indirectly.  It needs to be included in both places,
unfortunately.  I don't see a wya around this.

Also note the potential functionality leakage -- I added an instance
method for #empty even though I only needed #empty?.  For that matter,
all my stubs have attr_writers that they probably shouldn't have.
There's a risk that CartView will accidentally acquire a dependency on
one of these (by a typo in this case) and it would not be added to the
CartView Cart interface, and therefore not be enforced on the real Cart.

Wow, this is a lot of files!

Should I be having a separate interface for nearly every
class-collaborator combo?  Would it be enough to forgo the Priceable
interface in favor of having only an Item interface that is stubbable
but would permit unintended dependencies on other parts of the Item API
(besides price, in this case)?

Can I automate the production of a stub, interface, and stub spec by
specifying all three at once?  They seem very predictable so far.