Extract CartView's item template as a collaborator and inject it
Because CartView is responsible for rendering a (potentially deeply)
nested Cart structure with items, the cart_view_spec must setup a
collaborating Cart with that complexity.  I would like to limit the
complexity of view spec setup.

Extract an inner CartItemView.  Allow this view to be injected into the
CartView so that we can test the CartView in isolation from
CartItemView.  This means we can pass a stubbed structure into the
CartView for testing purposes, which the CartItemViewStub will
know how to render appropriately.  Create the requisite interface to
ensure compatibility between the CartItemView and CartItemViewStub, and
also create the requisite specs.

Wow, this is a lot of files!  (Six new in this commit, 9 changed all
together.)  And yet, as the Cart structure grows in complexity, number
of test permutations, and setup cost, it seems critical to be able to
cleave it away from other templates for testing purposes.